+++ internal/controller/operator/factory/vmdistributed/zone.go
@@ -30,10 +30,11 @@ import (
 )
 
 type zones struct {
-	httpClient *http.Client
-	vmagents   []*vmv1beta1.VMAgent
-	vmclusters []*vmv1beta1.VMCluster
-	hasChanges []bool
+	httpClient   *http.Client
+	vmagents     []*vmv1beta1.VMAgent
+	vmclusters   []*vmv1beta1.VMCluster
+	hasChanges   []bool
+	trafficModes []vmv1alpha1.VMDistributedTrafficMode
 }
 
 func (zs *zones) Len() int {
@@ -66,6 +67,7 @@ func (zs *zones) Swap(i, j int) {
 	zs.vmagents[i], zs.vmagents[j] = zs.vmagents[j], zs.vmagents[i]
 	zs.vmclusters[i], zs.vmclusters[j] = zs.vmclusters[j], zs.vmclusters[i]
 	zs.hasChanges[i], zs.hasChanges[j] = zs.hasChanges[j], zs.hasChanges[i]
+	zs.trafficModes[i], zs.trafficModes[j] = zs.trafficModes[j], zs.trafficModes[i]
 }
 
 // getZones builds desired zones
@@ -74,9 +76,10 @@ func getZones(ctx context.Context, rclient client.Client, cr *vmv1alpha1.VMDistr
 		httpClient: &http.Client{
 			Timeout: httpTimeout,
 		},
-		vmagents:   make([]*vmv1beta1.VMAgent, len(cr.Spec.Zones)),
-		vmclusters: make([]*vmv1beta1.VMCluster, len(cr.Spec.Zones)),
-		hasChanges: make([]bool, len(cr.Spec.Zones)),
+		vmagents:     make([]*vmv1beta1.VMAgent, len(cr.Spec.Zones)),
+		vmclusters:   make([]*vmv1beta1.VMCluster, len(cr.Spec.Zones)),
+		hasChanges:   make([]bool, len(cr.Spec.Zones)),
+		trafficModes: make([]vmv1alpha1.VMDistributedTrafficMode, len(cr.Spec.Zones)),
 	}
 	for i := range cr.Spec.Zones {
 		z := &cr.Spec.Zones[i]
@@ -102,8 +105,12 @@ func getZones(ctx context.Context, rclient client.Client, cr *vmv1alpha1.VMDistr
 		prevClusterSpec := vmCluster.Spec
 		vmCluster.Spec = *vmClusterSpec
 		rclient.Scheme().Default(&vmCluster)
+		if (z.TrafficMode == vmv1alpha1.VMDistributedTrafficModeReadOnly || z.TrafficMode == vmv1alpha1.VMDistributedTrafficModeMaintenance) && vmCluster.Spec.VMInsert != nil {
+			vmCluster.Spec.VMInsert.ReplicaCount = ptr.To(int32(0))
+		}
 		zs.hasChanges[i] = !equality.Semantic.DeepEqual(&vmCluster.Spec, &prevClusterSpec)
 		zs.vmclusters[i] = &vmCluster
+		zs.trafficModes[i] = z.TrafficMode
 	}
 
 	for i := range cr.Spec.Zones {
@@ -238,7 +245,7 @@ func (zs *zones) updateLB(ctx context.Context, rclient client.Client, cr *vmv1al
 	if !ptr.Deref(cr.Spec.VMAuth.Enabled, true) {
 		return nil
 	}
-	vmAuth := buildVMAuthLB(cr, zs.vmagents, zs.vmclusters, excludeIds...)
+	vmAuth := buildVMAuthLB(cr, zs.vmagents, zs.vmclusters, zs.trafficModes, excludeIds...)
 	if vmAuth == nil {
 		return nil
 	}